refactor(tasks): move Go2W rough owner [skip ci]
