+++ src/unilab/envs/locomotion/go2w/__init__.py
@@ -1,2 +1,1 @@
 from .joystick import Go2WJoystickCfg, Go2WJoystickEnv
-from .rough import Go2WJoystickRoughCfg, Go2WJoystickRoughEnv
