+++ src/unilab/tasks/locomotion/go2w/__init__.py
@@ -1,0 +1,3 @@
+from .rough import Go2WJoystickRoughCfg, Go2WJoystickRoughEnv
+
+__all__ = ["Go2WJoystickRoughCfg", "Go2WJoystickRoughEnv"]
