+++ tests/envs/locomotion/test_go2_terrain_spawn.py
@@ -116,7 +116,7 @@ def test_go1_rough_initialization_and_reset_use_backend_terrain_contract():
 
 def test_go2w_rough_initialization_and_reset_use_backend_terrain_contract():
     from unilab.envs.locomotion.go2w.joystick import RewardConfig
-    from unilab.envs.locomotion.go2w.rough import Go2WJoystickRoughCfg, Go2WJoystickRoughEnv
+    from unilab.tasks.locomotion.go2w.rough import Go2WJoystickRoughCfg, Go2WJoystickRoughEnv
 
     cfg = Go2WJoystickRoughCfg(
         reward_config=RewardConfig(scales={}, tracking_sigma=0.25, base_height_target=0.3)
