+++ tests/tasks/test_package_boundary.py
@@ -15,6 +15,7 @@ _TASK_REGISTRY_MODULES = (
     "unilab.tasks.locomotion.go1",
     "unilab.tasks.locomotion.go2",
     "unilab.envs.locomotion.go2w",
+    "unilab.tasks.locomotion.go2w.rough",
     "unilab.envs.locomotion.g1",
     "unilab.envs.locomotion.go2_arm",
     "unilab.tasks.locomotion.a2",
